refactor(harness): extract Pipeline background from file route
Moves the Pipeline background, empty-state onboarding, and shared home content out of the
`/` TanStack file route. Settings and Session Telemetry now import the same non-route
`PipelinePage`, and board/repository consumers use the extracted shared module so the
dependency graph no longer reaches `routes/index.tsx`.

This restores route code splitting while preserving Repos’ shared blank-slate/onboarding
behavior and existing routed-overlay navigation.

Verification: `harness:lint`, `harness:typecheck`, `harness:build`, and
`harness:test` all passed. The production build no longer reports the TanStack Router
`PipelinePage` code-splitting warning.

Closes #861
