+++ apps/harness/test/agent-backend-warning-contrast.test.tsx
@@ -29,7 +29,7 @@ const rootSource = () =>
   readFileSync(join(import.meta.dir, "../src/routes/__root.tsx"), "utf8")
 
 const indexSource = () =>
-  readFileSync(join(import.meta.dir, "../src/routes/index.tsx"), "utf8")
+  readFileSync(join(import.meta.dir, "../src/home-page-content.tsx"), "utf8")
 
 const warningsMarkup = () =>
   renderToStaticMarkup(<AgentBackendWarnings warnings={[BEDROCK_WARNING]} />)
