+++ apps/harness/test/interchange-phase5.test.ts
@@ -6,7 +6,7 @@ const rootSource = () =>
   readFileSync(join(import.meta.dir, "../src/routes/__root.tsx"), "utf8")
 
 const homeSource = () =>
-  readFileSync(join(import.meta.dir, "../src/routes/index.tsx"), "utf8")
+  readFileSync(join(import.meta.dir, "../src/home-page-content.tsx"), "utf8")
 
 const uiSource = () =>
   readFileSync(join(import.meta.dir, "../src/ui.ts"), "utf8")
