+++ apps/harness/test/jobs-reset-visibility.test.ts
@@ -4,7 +4,7 @@ import { canShowWorkItemResetAction } from "../src/work-item-job-actions.js"
 import { describe, expect, test } from "bun:test"
 
 const homeSource = () =>
-  readFileSync(join(import.meta.dir, "../src/routes/index.tsx"), "utf8")
+  readFileSync(join(import.meta.dir, "../src/home-page-content.tsx"), "utf8")
 
 const lifecycleStatusSource = (): string => {
   const source = homeSource()
