+++ apps/harness/test/repository-settings-agent-backend.test.ts
@@ -3,7 +3,7 @@ import { join } from "node:path"
 import { describe, expect, test } from "bun:test"
 
 const indexSource = () =>
-  readFileSync(join(import.meta.dir, "../src/routes/index.tsx"), "utf8")
+  readFileSync(join(import.meta.dir, "../src/home-page-content.tsx"), "utf8")
 
 describe("Repository settings Agent Backend override", () => {
   test("offers harness-default option and places backend control above models", () => {
